added 3 hooks for classDetails, resources, and tasks, all of which take in classId string, integrated with class/[classId]/page.tsx
